Allow AWS IDs and region to be overridden
Work in support of automated AWS deploy
https://github.com/ChurchApps/ChurchAppsSupport/issues/507

- Use $AWS_REGION from the environment.
- Dynamically find the AWS account ID. Replaces
  hard-coded account ID.
- Read $SERVERLESS_LAYERS_BUCKET,
  $SECURITY_GROUP_IDS, and $SUBNET_IDS from
  the environment. The latter two can be a single
  ID or a comma-separated list of IDs.

`variablesResolutionMode: 20210326` was added to
make the `${aws:accountId}` lookup work.

I could not find any docs for serverless v2, and
LLMs were not helpful. I had to resort to trial
and error to find a working method of adding
configurable values. Env vars are the only thing
I found that worked.